Trait rusoto_workmail::Workmail

pub trait Workmail {
    fn associate_delegate_to_resource(
        &self, 
        input: AssociateDelegateToResourceRequest
    ) -> RusotoFuture<AssociateDelegateToResourceResponse, AssociateDelegateToResourceError>;

    fn associate_member_to_group(
        &self, 
        input: AssociateMemberToGroupRequest
    ) -> RusotoFuture<AssociateMemberToGroupResponse, AssociateMemberToGroupError>;

    fn create_alias(
        &self, 
        input: CreateAliasRequest
    ) -> RusotoFuture<CreateAliasResponse, CreateAliasError>;

    fn create_group(
        &self, 
        input: CreateGroupRequest
    ) -> RusotoFuture<CreateGroupResponse, CreateGroupError>;

    fn create_resource(
        &self, 
        input: CreateResourceRequest
    ) -> RusotoFuture<CreateResourceResponse, CreateResourceError>;

    fn create_user(
        &self, 
        input: CreateUserRequest
    ) -> RusotoFuture<CreateUserResponse, CreateUserError>;

    fn delete_alias(
        &self, 
        input: DeleteAliasRequest
    ) -> RusotoFuture<DeleteAliasResponse, DeleteAliasError>;

    fn delete_group(
        &self, 
        input: DeleteGroupRequest
    ) -> RusotoFuture<DeleteGroupResponse, DeleteGroupError>;

    fn delete_mailbox_permissions(
        &self, 
        input: DeleteMailboxPermissionsRequest
    ) -> RusotoFuture<DeleteMailboxPermissionsResponse, DeleteMailboxPermissionsError>;

    fn delete_resource(
        &self, 
        input: DeleteResourceRequest
    ) -> RusotoFuture<DeleteResourceResponse, DeleteResourceError>;

    fn delete_user(
        &self, 
        input: DeleteUserRequest
    ) -> RusotoFuture<DeleteUserResponse, DeleteUserError>;

    fn deregister_from_work_mail(
        &self, 
        input: DeregisterFromWorkMailRequest
    ) -> RusotoFuture<DeregisterFromWorkMailResponse, DeregisterFromWorkMailError>;

    fn describe_group(
        &self, 
        input: DescribeGroupRequest
    ) -> RusotoFuture<DescribeGroupResponse, DescribeGroupError>;

    fn describe_organization(
        &self, 
        input: DescribeOrganizationRequest
    ) -> RusotoFuture<DescribeOrganizationResponse, DescribeOrganizationError>;

    fn describe_resource(
        &self, 
        input: DescribeResourceRequest
    ) -> RusotoFuture<DescribeResourceResponse, DescribeResourceError>;

    fn describe_user(
        &self, 
        input: DescribeUserRequest
    ) -> RusotoFuture<DescribeUserResponse, DescribeUserError>;

    fn disassociate_delegate_from_resource(
        &self, 
        input: DisassociateDelegateFromResourceRequest
    ) -> RusotoFuture<DisassociateDelegateFromResourceResponse, DisassociateDelegateFromResourceError>;

    fn disassociate_member_from_group(
        &self, 
        input: DisassociateMemberFromGroupRequest
    ) -> RusotoFuture<DisassociateMemberFromGroupResponse, DisassociateMemberFromGroupError>;

    fn get_mailbox_details(
        &self, 
        input: GetMailboxDetailsRequest
    ) -> RusotoFuture<GetMailboxDetailsResponse, GetMailboxDetailsError>;

    fn list_aliases(
        &self, 
        input: ListAliasesRequest
    ) -> RusotoFuture<ListAliasesResponse, ListAliasesError>;

    fn list_group_members(
        &self, 
        input: ListGroupMembersRequest
    ) -> RusotoFuture<ListGroupMembersResponse, ListGroupMembersError>;

    fn list_groups(
        &self, 
        input: ListGroupsRequest
    ) -> RusotoFuture<ListGroupsResponse, ListGroupsError>;

    fn list_mailbox_permissions(
        &self, 
        input: ListMailboxPermissionsRequest
    ) -> RusotoFuture<ListMailboxPermissionsResponse, ListMailboxPermissionsError>;

    fn list_organizations(
        &self, 
        input: ListOrganizationsRequest
    ) -> RusotoFuture<ListOrganizationsResponse, ListOrganizationsError>;

    fn list_resource_delegates(
        &self, 
        input: ListResourceDelegatesRequest
    ) -> RusotoFuture<ListResourceDelegatesResponse, ListResourceDelegatesError>;

    fn list_resources(
        &self, 
        input: ListResourcesRequest
    ) -> RusotoFuture<ListResourcesResponse, ListResourcesError>;

    fn list_users(
        &self, 
        input: ListUsersRequest
    ) -> RusotoFuture<ListUsersResponse, ListUsersError>;

    fn put_mailbox_permissions(
        &self, 
        input: PutMailboxPermissionsRequest
    ) -> RusotoFuture<PutMailboxPermissionsResponse, PutMailboxPermissionsError>;

    fn register_to_work_mail(
        &self, 
        input: RegisterToWorkMailRequest
    ) -> RusotoFuture<RegisterToWorkMailResponse, RegisterToWorkMailError>;

    fn reset_password(
        &self, 
        input: ResetPasswordRequest
    ) -> RusotoFuture<ResetPasswordResponse, ResetPasswordError>;

    fn update_mailbox_quota(
        &self, 
        input: UpdateMailboxQuotaRequest
    ) -> RusotoFuture<UpdateMailboxQuotaResponse, UpdateMailboxQuotaError>;

    fn update_primary_email_address(
        &self, 
        input: UpdatePrimaryEmailAddressRequest
    ) -> RusotoFuture<UpdatePrimaryEmailAddressResponse, UpdatePrimaryEmailAddressError>;

    fn update_resource(
        &self, 
        input: UpdateResourceRequest
    ) -> RusotoFuture<UpdateResourceResponse, UpdateResourceError>;
}

Trait representing the capabilities of the Amazon WorkMail API. Amazon WorkMail clients implement this trait.

Required methods

fn associate_delegate_to_resource(
&self,
input: AssociateDelegateToResourceRequest
) -> RusotoFuture<AssociateDelegateToResourceResponse, AssociateDelegateToResourceError>

Adds a member (user or group) to the resource's set of delegates.

fn associate_member_to_group(
&self,
input: AssociateMemberToGroupRequest
) -> RusotoFuture<AssociateMemberToGroupResponse, AssociateMemberToGroupError>

Adds a member (user or group) to the group's set.

fn create_alias(
&self,
input: CreateAliasRequest
) -> RusotoFuture<CreateAliasResponse, CreateAliasError>

Adds an alias to the set of a given member (user or group) of Amazon WorkMail.

fn create_group(
&self,
input: CreateGroupRequest
) -> RusotoFuture<CreateGroupResponse, CreateGroupError>

Creates a group that can be used in Amazon WorkMail by calling the RegisterToWorkMail operation.

fn create_resource(
&self,
input: CreateResourceRequest
) -> RusotoFuture<CreateResourceResponse, CreateResourceError>

Creates a new Amazon WorkMail resource.

fn create_user(
&self,
input: CreateUserRequest
) -> RusotoFuture<CreateUserResponse, CreateUserError>

Creates a user who can be used in Amazon WorkMail by calling the RegisterToWorkMail operation.

fn delete_alias(
&self,
input: DeleteAliasRequest
) -> RusotoFuture<DeleteAliasResponse, DeleteAliasError>

Remove one or more specified aliases from a set of aliases for a given user.

fn delete_group(
&self,
input: DeleteGroupRequest
) -> RusotoFuture<DeleteGroupResponse, DeleteGroupError>

Deletes a group from Amazon WorkMail.

fn delete_mailbox_permissions(
&self,
input: DeleteMailboxPermissionsRequest
) -> RusotoFuture<DeleteMailboxPermissionsResponse, DeleteMailboxPermissionsError>

Deletes permissions granted to a member (user or group).

fn delete_resource(
&self,
input: DeleteResourceRequest
) -> RusotoFuture<DeleteResourceResponse, DeleteResourceError>

Deletes the specified resource.

fn delete_user(
&self,
input: DeleteUserRequest
) -> RusotoFuture<DeleteUserResponse, DeleteUserError>

Deletes a user from Amazon WorkMail and all subsequent systems. Before you can delete a user, the user state must be DISABLED. Use the DescribeUser action to confirm the user state.

Deleting a user is permanent and cannot be undone. WorkMail archives user mailboxes for 30 days before they are permanently removed.

fn deregister_from_work_mail(
&self,
input: DeregisterFromWorkMailRequest
) -> RusotoFuture<DeregisterFromWorkMailResponse, DeregisterFromWorkMailError>

Mark a user, group, or resource as no longer used in Amazon WorkMail. This action disassociates the mailbox and schedules it for clean-up. WorkMail keeps mailboxes for 30 days before they are permanently removed. The functionality in the console is Disable.

fn describe_group(
&self,
input: DescribeGroupRequest
) -> RusotoFuture<DescribeGroupResponse, DescribeGroupError>

Returns the data available for the group.

fn describe_organization(
&self,
input: DescribeOrganizationRequest
) -> RusotoFuture<DescribeOrganizationResponse, DescribeOrganizationError>

Provides more information regarding a given organization based on its identifier.

fn describe_resource(
&self,
input: DescribeResourceRequest
) -> RusotoFuture<DescribeResourceResponse, DescribeResourceError>

Returns the data available for the resource.

fn describe_user(
&self,
input: DescribeUserRequest
) -> RusotoFuture<DescribeUserResponse, DescribeUserError>

Provides information regarding the user.

fn disassociate_delegate_from_resource(
&self,
input: DisassociateDelegateFromResourceRequest
) -> RusotoFuture<DisassociateDelegateFromResourceResponse, DisassociateDelegateFromResourceError>

Removes a member from the resource's set of delegates.

fn disassociate_member_from_group(
&self,
input: DisassociateMemberFromGroupRequest
) -> RusotoFuture<DisassociateMemberFromGroupResponse, DisassociateMemberFromGroupError>

Removes a member from a group.

fn get_mailbox_details(
&self,
input: GetMailboxDetailsRequest
) -> RusotoFuture<GetMailboxDetailsResponse, GetMailboxDetailsError>

Requests a user's mailbox details for a specified organization and user.

fn list_aliases(
&self,
input: ListAliasesRequest
) -> RusotoFuture<ListAliasesResponse, ListAliasesError>

Creates a paginated call to list the aliases associated with a given entity.

fn list_group_members(
&self,
input: ListGroupMembersRequest
) -> RusotoFuture<ListGroupMembersResponse, ListGroupMembersError>

Returns an overview of the members of a group. Users and groups can be members of a group.

fn list_groups(
&self,
input: ListGroupsRequest
) -> RusotoFuture<ListGroupsResponse, ListGroupsError>

Returns summaries of the organization's groups.

fn list_mailbox_permissions(
&self,
input: ListMailboxPermissionsRequest
) -> RusotoFuture<ListMailboxPermissionsResponse, ListMailboxPermissionsError>

Lists the mailbox permissions associated with a user, group, or resource mailbox.

fn list_organizations(
&self,
input: ListOrganizationsRequest
) -> RusotoFuture<ListOrganizationsResponse, ListOrganizationsError>

Returns summaries of the customer's non-deleted organizations.

fn list_resource_delegates(
&self,
input: ListResourceDelegatesRequest
) -> RusotoFuture<ListResourceDelegatesResponse, ListResourceDelegatesError>

Lists the delegates associated with a resource. Users and groups can be resource delegates and answer requests on behalf of the resource.

fn list_resources(
&self,
input: ListResourcesRequest
) -> RusotoFuture<ListResourcesResponse, ListResourcesError>

Returns summaries of the organization's resources.

fn list_users(
&self,
input: ListUsersRequest
) -> RusotoFuture<ListUsersResponse, ListUsersError>

Returns summaries of the organization's users.

fn put_mailbox_permissions(
&self,
input: PutMailboxPermissionsRequest
) -> RusotoFuture<PutMailboxPermissionsResponse, PutMailboxPermissionsError>

Sets permissions for a user, group, or resource. This replaces any pre-existing permissions.

fn register_to_work_mail(
&self,
input: RegisterToWorkMailRequest
) -> RusotoFuture<RegisterToWorkMailResponse, RegisterToWorkMailError>

Registers an existing and disabled user, group, or resource for Amazon WorkMail use by associating a mailbox and calendaring capabilities. It performs no change if the user, group, or resource is enabled and fails if the user, group, or resource is deleted. This operation results in the accumulation of costs. For more information, see Pricing. The equivalent console functionality for this operation is Enable.

Users can either be created by calling the CreateUser API operation or they can be synchronized from your directory. For more information, see DeregisterFromWorkMail.

fn reset_password(
&self,
input: ResetPasswordRequest
) -> RusotoFuture<ResetPasswordResponse, ResetPasswordError>

Allows the administrator to reset the password for a user.

fn update_mailbox_quota(
&self,
input: UpdateMailboxQuotaRequest
) -> RusotoFuture<UpdateMailboxQuotaResponse, UpdateMailboxQuotaError>

Updates a user's current mailbox quota for a specified organization and user.

fn update_primary_email_address(
&self,
input: UpdatePrimaryEmailAddressRequest
) -> RusotoFuture<UpdatePrimaryEmailAddressResponse, UpdatePrimaryEmailAddressError>

Updates the primary email for a user, group, or resource. The current email is moved into the list of aliases (or swapped between an existing alias and the current primary email), and the email provided in the input is promoted as the primary.

fn update_resource(
&self,
input: UpdateResourceRequest
) -> RusotoFuture<UpdateResourceResponse, UpdateResourceError>

Updates data for the resource. To have the latest information, it must be preceded by a DescribeResource call. The dataset in the request should be the one expected when performing another DescribeResource call.

Loading content...