Crate rusoto_workmail

Amazon WorkMail is a secure, managed business email and calendaring service with support for existing desktop and mobile email clients. You can access your email, contacts, and calendars using Microsoft Outlook, your browser, or other native iOS and Android email applications. You can integrate WorkMail with your existing corporate directory and control both the keys that encrypt your data and the location in which your data is stored.

The WorkMail API is designed for the following scenarios:

Listing and describing organizations

Managing users

Managing groups

Managing resources

All WorkMail API operations are Amazon-authenticated and certificate-signed. They not only require the use of the AWS SDK, but also allow for the exclusive use of AWS Identity and Access Management users and roles to help facilitate access, trust, and permission policies. By creating a role and allowing an IAM user to access the WorkMail site, the IAM user gains full administrative visibility into the entire WorkMail organization (or as set in the IAM policy). This includes, but is not limited to, the ability to create, update, and delete users, groups, and resources. This allows developers to perform the scenarios listed above, as well as give users the ability to grant access on a selective basis using the IAM model.

If you're using the service, you're probably looking for WorkmailClient and Workmail.

Structs

AssociateDelegateToResourceRequest

AssociateDelegateToResourceResponse

AssociateMemberToGroupRequest

AssociateMemberToGroupResponse

BookingOptions
At least one delegate must be associated to the resource to disable automatic replies from the resource.

CreateAliasRequest

CreateAliasResponse

CreateGroupRequest

CreateGroupResponse

CreateResourceRequest

CreateResourceResponse

CreateUserRequest

CreateUserResponse

Delegate
The name of the attribute, which is one of the values defined in the UserAttribute enumeration.

DeleteAliasRequest

DeleteAliasResponse

DeleteGroupRequest

DeleteGroupResponse

DeleteMailboxPermissionsRequest

DeleteMailboxPermissionsResponse

DeleteResourceRequest

DeleteResourceResponse

DeleteUserRequest

DeleteUserResponse

DeregisterFromWorkMailRequest

DeregisterFromWorkMailResponse

DescribeGroupRequest

DescribeGroupResponse

DescribeOrganizationRequest

DescribeOrganizationResponse

DescribeResourceRequest

DescribeResourceResponse

DescribeUserRequest

DescribeUserResponse

DisassociateDelegateFromResourceRequest

DisassociateDelegateFromResourceResponse

DisassociateMemberFromGroupRequest

DisassociateMemberFromGroupResponse

GetMailboxDetailsRequest

GetMailboxDetailsResponse

Group
The representation of an Amazon WorkMail group.

ListAliasesRequest

ListAliasesResponse

ListGroupMembersRequest

ListGroupMembersResponse

ListGroupsRequest

ListGroupsResponse

ListMailboxPermissionsRequest

ListMailboxPermissionsResponse

ListOrganizationsRequest

ListOrganizationsResponse

ListResourceDelegatesRequest

ListResourceDelegatesResponse

ListResourcesRequest

ListResourcesResponse

ListUsersRequest

ListUsersResponse

Member
The representation of a user or group.

OrganizationSummary
The representation of an organization.

Permission
Permission granted to a user, group, or resource to access a certain aspect of another user, group, or resource mailbox.

PutMailboxPermissionsRequest

PutMailboxPermissionsResponse

RegisterToWorkMailRequest

RegisterToWorkMailResponse

ResetPasswordRequest

ResetPasswordResponse

Resource
The representation of a resource.

UpdateMailboxQuotaRequest

UpdateMailboxQuotaResponse

UpdatePrimaryEmailAddressRequest

UpdatePrimaryEmailAddressResponse

UpdateResourceRequest

UpdateResourceResponse

User
The representation of an Amazon WorkMail user.

WorkmailClient
A client for the Amazon WorkMail API.

Enums

AssociateDelegateToResourceError
Errors returned by AssociateDelegateToResource

AssociateMemberToGroupError
Errors returned by AssociateMemberToGroup

CreateAliasError
Errors returned by CreateAlias

CreateGroupError
Errors returned by CreateGroup

CreateResourceError
Errors returned by CreateResource

CreateUserError
Errors returned by CreateUser

DeleteAliasError
Errors returned by DeleteAlias

DeleteGroupError
Errors returned by DeleteGroup

DeleteMailboxPermissionsError
Errors returned by DeleteMailboxPermissions

DeleteResourceError
Errors returned by DeleteResource

DeleteUserError
Errors returned by DeleteUser

DeregisterFromWorkMailError
Errors returned by DeregisterFromWorkMail

DescribeGroupError
Errors returned by DescribeGroup

DescribeOrganizationError
Errors returned by DescribeOrganization

DescribeResourceError
Errors returned by DescribeResource

DescribeUserError
Errors returned by DescribeUser

DisassociateDelegateFromResourceError
Errors returned by DisassociateDelegateFromResource

DisassociateMemberFromGroupError
Errors returned by DisassociateMemberFromGroup

GetMailboxDetailsError
Errors returned by GetMailboxDetails

ListAliasesError
Errors returned by ListAliases

ListGroupMembersError
Errors returned by ListGroupMembers

ListGroupsError
Errors returned by ListGroups

ListMailboxPermissionsError
Errors returned by ListMailboxPermissions

ListOrganizationsError
Errors returned by ListOrganizations

ListResourceDelegatesError
Errors returned by ListResourceDelegates

ListResourcesError
Errors returned by ListResources

ListUsersError
Errors returned by ListUsers

PutMailboxPermissionsError
Errors returned by PutMailboxPermissions

RegisterToWorkMailError
Errors returned by RegisterToWorkMail

ResetPasswordError
Errors returned by ResetPassword

UpdateMailboxQuotaError
Errors returned by UpdateMailboxQuota

UpdatePrimaryEmailAddressError
Errors returned by UpdatePrimaryEmailAddress

UpdateResourceError
Errors returned by UpdateResource

Traits

Workmail
Trait representing the capabilities of the Amazon WorkMail API. Amazon WorkMail clients implement this trait.