Struct rusoto_workmail::WorkmailClient

pub struct WorkmailClient { /* fields omitted */ }

A client for the Amazon WorkMail API.

Methods

impl WorkmailClient

pub fn new(region: Region) -> WorkmailClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> WorkmailClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,